backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
elasticsearch[helpers]>=8.11.0,<9.0.0
celery>=5.3.6
redis>=5.0.1
sentence-transformers>=2.7.0
langchain>=0.1.16
pypdf2>=3.0.1
python-docx>=1.1.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
python-multipart>=0.0.6
google-cloud-logging>=3.5.0
google-cloud-secret-manager>=2.18.0
